fix: since the dropdown menu in the dataset page was a little bit hidden, switched dropdown menu to buttons
