+++ .changeset/slow-swans-cough.md
@@ -1,0 +1,5 @@
+---
+"geohub": patch
+---
+
+fix: since the dropdown menu in the dataset page was a little bit hidden, switched dropdown menu to buttons
